pyproject.toml
[tool.poetry]
name = "personal-assistant-mcp"
version = "0.1.0"
description = "Personal assistant MCP server"
authors = [
    "Oleksandr Hryhor <lemannrus@gmail.com>"
]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
mcp = ">=1.6.0,<2.0.0"
google-auth = "^2.38.0"
google-api-python-client = "^2.167.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
beautifulsoup4 = "^4.13.3"
py-trello = "^0.20.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
